Time: 0 ms (100.00%) | Memory: 7.8 MB (79.36%) - LeetSync

+++ 3379-score-of-a-string/score-of-a-string.cpp
@@ -3,7 +3,7 @@ public:
     int scoreOfString(string s) {
         long long sum=0;
         for(int i=0 ; i<s.size()-1 ; i++){
-            sum += abs((int)s[i]-(int)s[i+1]);
+            sum += abs(s[i]-s[i+1]);
         }
         return sum;
     }
